+++ src/components/BaseColumn.php
@@ -467,7 +467,18 @@ abstract class BaseColumn extends BaseObject
      */
     protected function renderDragColumn($name, $value, $options)
     {
-        return Html::tag('span', null, ['class' => 'glyphicon glyphicon-menu-hamburger drag-handle']);
+        // return Html::tag('span', null, ['class' => 'glyphicon glyphicon-menu-hamburger drag-handle']);
+        /**
+         * Class was passed into options by TableRendered->renderCellContent(),
+         * we can extract it here
+         */
+        $class = '';
+        if (array_key_exists('class', $options)) {
+            $class = ArrayHelper::remove($options, 'class');
+        }
+        $dragClass = implode(" ", [$class, 'drag-handle']);
+
+        return Html::tag('span', null, ['class'=>$dragClass]);
     }
 
     /**