+++ src/renderers/BaseRenderer.php
@@ -248,7 +248,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->removeButtonOptions)) {
-            $this->removeButtonOptions['label'] = Html::tag('i', null, ['class' => 'glyphicon glyphicon-remove']);
+            $this->removeButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('remove')]);
         }
 
         if (!array_key_exists('class', $this->addButtonOptions)) {
@@ -256,7 +256,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->addButtonOptions)) {
-            $this->addButtonOptions['label'] = Html::tag('i', null, ['class' => 'glyphicon glyphicon-plus']);
+            $this->addButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('add')]);
         }
 
         if (!array_key_exists('class', $this->cloneButtonOptions)) {
@@ -264,7 +264,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->cloneButtonOptions)) {
-            $this->cloneButtonOptions['label'] = Html::tag('i', null, ['class' => 'glyphicon glyphicon-duplicate']);
+            $this->cloneButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('clone')]);
         }
     }
 
@@ -488,5 +488,16 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         return $attributes;
+    }
+
+    /**
+     * @param $action - the control parameter, used as key into allowed types
+     * @return string - the relevant icon class
+     */
+    protected function getFontClass($action) {
+        if (in_array($action, ['add', 'remove', 'clone', 'drag-handle'])) {
+            return $this->fontMap[$action];
+        }
+        return '';
     }
 }